pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
highrise-bot-sdk="23.3.4"
ffmpeg = "^1.4"
flask = "^3.0.3"
youtube-dl = "^2021.12.17"
google-api-python-client = "^2.127.0"
pafy = "^0.5.5"
pyfiglet = "^1.0.2"
python-dotenv = "^1.0.1"
python-vlc = "^3.0.20123"
pyglet = "^2.0.15"
pytube = "15.0.0"
youtube-search-python = "^1.6.6"
pydub = "^0.25.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"